Stef And Petr changes
