+++ src/main/java/com/jagrosh/jmusicbot/commands/music/PlayCmd.java
@@ -30,8 +30,7 @@ import com.jagrosh.jmusicbot.audio.AudioHandler;
 import com.jagrosh.jmusicbot.audio.QueuedTrack;
 import com.jagrosh.jmusicbot.commands.DJCommand;
 import com.jagrosh.jmusicbot.commands.MusicCommand;
-import com.jagrosh.jmusicbot.playlist.PlaylistLoader;
-import com.jagrosh.jmusicbot.playlist.Playlist;
+import com.jagrosh.jmusicbot.playlist.PlaylistLoader.Playlist;
 import com.jagrosh.jmusicbot.utils.FormatUtil;
 import java.util.concurrent.TimeUnit;
 import net.dv8tion.jda.api.Permission;
@@ -48,6 +47,7 @@ public class PlayCmd extends MusicCommand
     private final static String CANCEL = "\uD83D\uDEAB"; // 🚫
     
     private final String loadingEmoji;
+    private final Message m;
     
     public PlayCmd(Bot bot)
     {
@@ -59,7 +59,7 @@ public class PlayCmd extends MusicCommand
         this.aliases = bot.getConfig().getAliases(this.name);
         this.beListening = true;
         this.bePlaying = false;
-        this.children = new Command[]{new PlaylistCmd(bot)};
+        this.children = new Command[]{new PlayCmdHelper(bot)};
     }
 
     @Override
@@ -91,184 +91,7 @@ public class PlayCmd extends MusicCommand
                 ? event.getArgs().substring(1,event.getArgs().length()-1) 
                 : event.getArgs().isEmpty() ? event.getMessage().getAttachments().get(0).getUrl() : event.getArgs();
         event.reply(loadingEmoji+" Loading... `["+args+"]`", m -> bot.getPlayerManager().loadItemOrdered(event.getGuild(), args, new ResultHandler(m,event,false)));
+        new ResultHandler(bot, m, event, false); // this was changed
     }
-    
-    private class ResultHandler implements AudioLoadResultHandler
-    {
-        private final Message m;
-        private final CommandEvent event;
-        private final boolean ytsearch;
-        
-        private ResultHandler(Message m, CommandEvent event, boolean ytsearch)
-        {
-            this.m = m;
-            this.event = event;
-            this.ytsearch = ytsearch;
-        }
-        
-        private void loadSingle(AudioTrack track, AudioPlaylist playlist)
-        {
-            if(bot.getConfig().isTooLong(track))
-            {
-                m.editMessage(FormatUtil.filter(event.getClient().getWarning()+" This track (**"+track.getInfo().title+"**) is longer than the allowed maximum: `"
-                        + TimeUtil.formatTime(track.getDuration())+"` > `"+ TimeUtil.formatTime(bot.getConfig().getMaxSeconds()*1000)+"`")).queue();
-                return;
-            }
-            AudioHandler handler = (AudioHandler)event.getGuild().getAudioManager().getSendingHandler();
-            int pos = handler.addTrack(new QueuedTrack(track, RequestMetadata.fromResultHandler(track, event)))+1;
-            String addMsg = FormatUtil.filter(event.getClient().getSuccess()+" Added **"+track.getInfo().title
-                    +"** (`"+ TimeUtil.formatTime(track.getDuration())+"`) "+(pos==0?"to begin playing":" to the queue at position "+pos));
-            if(playlist==null || !event.getSelfMember().hasPermission(event.getTextChannel(), Permission.MESSAGE_ADD_REACTION))
-                m.editMessage(addMsg).queue();
-            else
-            {
-                new ButtonMenu.Builder()
-                        .setText(addMsg+"\n"+event.getClient().getWarning()+" This track has a playlist of **"+playlist.getTracks().size()+"** tracks attached. Select "+LOAD+" to load playlist.")
-                        .setChoices(LOAD, CANCEL)
-                        .setEventWaiter(bot.getWaiter())
-                        .setTimeout(30, TimeUnit.SECONDS)
-                        .setAction(re ->
-                        {
-                            if(re.getName().equals(LOAD))
-                                m.editMessage(addMsg+"\n"+event.getClient().getSuccess()+" Loaded **"+loadPlaylist(playlist, track)+"** additional tracks!").queue();
-                            else
-                                m.editMessage(addMsg).queue();
-                        }).setFinalAction(m ->
-                        {
-                            try{ m.clearReactions().queue(); }catch(PermissionException ignore) {}
-                        }).build().display(m);
-            }
-        }
-        
-        private int loadPlaylist(AudioPlaylist playlist, AudioTrack exclude)
-        {
-            int[] count = {0};
-            playlist.getTracks().stream().forEach((track) -> {
-                if(!bot.getConfig().isTooLong(track) && !track.equals(exclude))
-                {
-                    AudioHandler handler = (AudioHandler)event.getGuild().getAudioManager().getSendingHandler();
-                    handler.addTrack(new QueuedTrack(track, RequestMetadata.fromResultHandler(track, event)));
-                    count[0]++;
-                }
-            });
-            return count[0];
-        }
-        
-        @Override
-        public void trackLoaded(AudioTrack track)
-        {
-            loadSingle(track, null);
-        }
-
-        @Override
-        public void playlistLoaded(AudioPlaylist playlist)
-        {
-            if(playlist.getTracks().size()==1 || playlist.isSearchResult())
-            {
-                AudioTrack single = playlist.getSelectedTrack()==null ? playlist.getTracks().get(0) : playlist.getSelectedTrack();
-                loadSingle(single, null);
-            }
-            else if (playlist.getSelectedTrack()!=null)
-            {
-                AudioTrack single = playlist.getSelectedTrack();
-                loadSingle(single, playlist);
-            }
-            else
-            {
-                int count = loadPlaylist(playlist, null);
-                if(playlist.getTracks().size() == 0)
-                {
-                    m.editMessage(FormatUtil.filter(event.getClient().getWarning()+" The playlist "+(playlist.getName()==null ? "" : "(**"+playlist.getName()
-                            +"**) ")+" could not be loaded or contained 0 entries")).queue();
-                }
-                else if(count==0)
-                {
-                    m.editMessage(FormatUtil.filter(event.getClient().getWarning()+" All entries in this playlist "+(playlist.getName()==null ? "" : "(**"+playlist.getName()
-                            +"**) ")+"were longer than the allowed maximum (`"+bot.getConfig().getMaxTime()+"`)")).queue();
-                }
-                else
-                {
-                    m.editMessage(FormatUtil.filter(event.getClient().getSuccess()+" Found "
-                            +(playlist.getName()==null?"a playlist":"playlist **"+playlist.getName()+"**")+" with `"
-                            + playlist.getTracks().size()+"` entries; added to the queue!"
-                            + (count<playlist.getTracks().size() ? "\n"+event.getClient().getWarning()+" Tracks longer than the allowed maximum (`"
-                            + bot.getConfig().getMaxTime()+"`) have been omitted." : ""))).queue();
-                }
-            }
-        }
-
-        @Override
-        public void noMatches()
-        {
-            if(ytsearch)
-                m.editMessage(FormatUtil.filter(event.getClient().getWarning()+" No results found for `"+event.getArgs()+"`.")).queue();
-            else
-                bot.getPlayerManager().loadItemOrdered(event.getGuild(), "ytsearch:"+event.getArgs(), new ResultHandler(m,event,true));
-        }
-
-        @Override
-        public void loadFailed(FriendlyException throwable)
-        {
-            if(throwable.severity==Severity.COMMON)
-                m.editMessage(event.getClient().getError()+" Error loading: "+throwable.getMessage()).queue();
-            else
-                m.editMessage(event.getClient().getError()+" Error loading track.").queue();
-        }
-    }
-    
-    public class PlaylistCmd extends MusicCommand
-    {
-        public PlaylistCmd(Bot bot)
-        {
-            super(bot);
-            this.name = "playlist";
-            this.aliases = new String[]{"pl"};
-            this.arguments = "<name>";
-            this.help = "plays the provided playlist";
-            this.beListening = true;
-            this.bePlaying = false;
-        }
-
-        @Override
-        public void doCommand(CommandEvent event) 
-        {
-            if(event.getArgs().isEmpty())
-            {
-                event.reply(event.getClient().getError()+" Please include a playlist name.");
-                return;
-            }
-            Playlist playlist = Playlist.loadPlaylist(event.getArgs());
-            if(playlist==null)
-            {
-                event.replyError("I could not find `"+event.getArgs()+".txt` in the Playlists folder.");
-                return;
-            }
-            event.getChannel().sendMessage(loadingEmoji+" Loading playlist **"+event.getArgs()+"**... ("+playlist.getTracks().size()+" items)").queue(m -> 
-            {
-                AudioHandler handler = (AudioHandler)event.getGuild().getAudioManager().getSendingHandler();
-                playlist.loadTracks(bot.getPlayerManager(), 
-                    (at) -> handler.addTrack(new QueuedTrack(at, RequestMetadata.fromResultHandler(at, event))), 
-                    () -> {
-                        StringBuilder builder = new StringBuilder(playlist.getTracks().isEmpty() 
-                                ? event.getClient().getWarning()+" No tracks were loaded!" 
-                                : event.getClient().getSuccess()+" Loaded **"+playlist.getTracks().size()+"** tracks!");
-                        
-                        if(!playlist.getErrors().isEmpty())
-                        {
-                            builder.append("\nThe following tracks failed to load:");
-                            playlist.getErrors().forEach(err -> 
-                                builder.append("\n`[").append(err.getIndex()+1).append("]` **")
-                                       .append(err.getItem()).append("**: ").append(err.getReason()));
-                        }
-                        
-                        String str = builder.toString();
-                        if(str.length() > 2000)
-                            str = str.substring(0, 1994) + " (...)";
-                        
-                        m.editMessage(FormatUtil.filter(str)).queue();
-                    }
-                );
-            });
-        }
-    }
+                
 }
