+++ src/main/java/com/jagrosh/jmusicbot/commands/music/PlayCmdHelper.java
