+++ src/main/java/com/jagrosh/jmusicbot/commands/music/ResultHandler.java
@@ -15,7 +15,7 @@ import com.sedmelluq.discord.lavaplayer.track.AudioPlaylist;
 import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Message;
-// import net.dv8tion.jda.api.exceptions.PermissionException;
+import net.dv8tion.jda.api.exceptions.PermissionException;
 
 import java.util.concurrent.TimeUnit;
