+++ src/main/java/com/jagrosh/jmusicbot/commands/owner/AppendlistCmd.java
@@ -63,3 +63,4 @@ import com.jagrosh.jmusicbot.playlist.PlaylistLoader.Playlist;
             }
         }
     }
+
