+++ src/main/java/com/jagrosh/jmusicbot/commands/owner/DeletelistCmd.java
@@ -1,0 +1,40 @@
+package com.jagrosh.jmusicbot.commands.owner;
+
+import java.io.IOException;
+
+import com.jagrosh.jmusicbot.Bot;
+import com.jagrosh.jmusicbot.commands.OwnerCommand;
+
+public class DeletelistCmd extends OwnerCommand 
+    {
+
+        public DeletelistCmd(Bot bot)
+        {
+            this.bot = bot;
+            this.name = "delete";
+            this.aliases = new String[]{"remove"};
+            this.help = "deletes an existing playlist";
+            this.arguments = "<name>";
+            this.guildOnly = false;
+        }
+
+        @Override
+        protected void execute(CommandEvent event) 
+        {
+            String pname = event.getArgs().replaceAll("\\s+", "_");
+            if(bot.getPlaylistLoader().getPlaylist(pname)==null)
+                event.reply(event.getClient().getError()+" Playlist `"+pname+"` doesn't exist!");
+            else
+            {
+                try
+                {
+                    bot.getPlaylistLoader().deletePlaylist(pname);
+                    event.reply(event.getClient().getSuccess()+" Successfully deleted playlist `"+pname+"`!");
+                }
+                catch(IOException e)
+                {
+                    event.reply(event.getClient().getError()+" I was unable to delete the playlist: "+e.getLocalizedMessage());
+                }
+            }
+        }
+    }
